+++ CesiumGltfContent/include/CesiumGltfContent/GltfUtilities.h
@@ -11,15 +11,16 @@
 #include <vector>
 
 namespace CesiumGltf {
+struct Buffer;
 struct Model;
-}
+} // namespace CesiumGltf
 
-namespace Cesium3DTilesContent {
+namespace CesiumGltfContent {
 /**
  * A collection of utility functions that are used to process and transform a
  * gltf model
  */
-struct CESIUM3DTILESCONTENT_API GltfUtilities {
+struct CESIUMGLTFCONTENT_API GltfUtilities {
   /**
    * @brief Applies the glTF's RTC_CENTER, if any, to the given transform.
    *
@@ -92,5 +93,33 @@ struct CESIUM3DTILESCONTENT_API GltfUtilities {
    */
   static std::vector<std::string_view>
   parseGltfCopyright(const CesiumGltf::Model& gltf);
+
+  /**
+   * @brief Merges all of the glTF's buffers into a single buffer (the first
+   * one).
+   *
+   * This is useful when writing the glTF as a GLB, which supports only a single
+   * embedded buffer.
+   *
+   * @param gltf The glTF in which to merge buffers.
+   */
+  static void collapseToSingleBuffer(CesiumGltf::Model& gltf);
+
+  /**
+   * @brief Copies the content of one {@link Buffer} to the end of another,
+   * updates all {@link BufferView} instances to refer to the destination
+   * buffer, and clears the contents of the original buffer.
+   *
+   * The source buffer is not removed, but it has a `byteLength` of zero after
+   * this function completes.
+   *
+   * @param gltf The glTF model to modify.
+   * @param destination The destination Buffer into which to move content.
+   * @param source The source Buffer from which to move content.
+   */
+  static void moveBufferContent(
+      CesiumGltf::Model& gltf,
+      CesiumGltf::Buffer& destination,
+      CesiumGltf::Buffer& source);
 };
-} // namespace Cesium3DTilesContent
+} // namespace CesiumGltfContent