+++ CesiumGltfContent/test/TestImageManipulation.cpp
@@ -1,13 +1,12 @@
-#include "CesiumGltfReader/ImageManipulation.h"
-
 #include <CesiumGltf/ImageCesium.h>
+#include <CesiumGltfContent/ImageManipulation.h>
 
 #include <catch2/catch.hpp>
 
 #include <algorithm>
 
 using namespace CesiumGltf;
-using namespace CesiumGltfReader;
+using namespace CesiumGltfContent;
 
 TEST_CASE("ImageManipulation::unsafeBlitImage entire image") {
   size_t width = 10;